pyproject.toml
[tool.poetry]
name = "ankigengpt"
version = "0.1.0"
description = ""
authors = ["rwxd <git@rwxd.de>"]
readme = "README.md"
packages = [{include = "ankigengpt"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
jinja2 = "^3.1.2"
openai = "^0.27.7"
pydub = "^0.25.1"
pyyaml = "^6.0"
beautifulsoup4 = "^4.12.2"
genanki = "^0.13.0"
tiktoken = "^0.4.0"
ebooklib = "^0.18"
rich = "^13.4.2"
pdfplumber = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[tool.poetry.scripts]
ankigengpt = "ankigengpt.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
